refactor: split must
